requirements.txt
app==0.0.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cssselect2==0.8.0
dnspython==2.8.0
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openpyxl==3.1.5
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pymongo==4.15.4
pyphen==0.17.2
python-dotenv==1.2.1
reportlab==4.4.5
requests==2.32.5
SQLAlchemy==2.0.44
stripe==14.0.0
tinycss2==1.5.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.4.0
gunicorn
psycopg2-binary==2.9.9

